backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
pymilvus
ollama
python-dotenv
pydantic-settings
pytest
pytest-asyncio
httpx
